+++ src/templates/index.ts
@@ -1,1 +1,1 @@
-export * from "./careers-item";
+export * from "./careers-single";
